+++ packages/wallet-management/src/index.ts
@@ -1,3 +1,3 @@
 export * from './LiFiWalletManagement';
 export * from './wallets';
-export * from './browserWallets';
+export * from './walletAutomation';
