+++ packages/wallet-management/src/walletAutomation.ts
